backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0      # ASGI server for FastAPI
python-multipart==0.0.6        # For file uploads

# AWS SDK
boto3==1.34.34                 # AWS services (DynamoDB, S3, Textract, SNS)

# Authentication
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4         # Password hashing
bcrypt<4.0.0                   # Compatible bcrypt version for passlib
python-dotenv==1.0.0           # Load .env files

# Utilities
pydantic==2.5.3               # Data validation
pydantic-settings==2.1.0      # Settings management
email-validator==2.3.0        # Email validation for Pydantic
Pillow==10.1.0                # Image processing


